pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chat-robot"
version = "0.1.0"
description = "Local chat robot with intent routing, RAG, and a single-GPU generation queue."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115,<1",
  "httpx>=0.28,<1",
  "uvicorn[standard]>=0.30,<1",
  "pydantic>=2.9,<3",
  "python-dotenv>=1.0,<2",
  "langsmith>=0.4.25,<1",
  "tomli>=2.0; python_version < '3.11'",
  "torch>=2.4,<3",
  "transformers>=4.56.2,<5",
  "accelerate>=1.13",
  "bitsandbytes>=0.49",
  "sentence-transformers>=5.1.2,<6",
  "sqlalchemy>=2.0,<3",
  "psycopg[binary]>=3.2,<4",
  "qdrant-client>=1.12,<2",
  "minio>=7.2,<8",
  "openpyxl>=3.1,<4",
  "python-docx>=1.1,<2",
  "python-multipart>=0.0.9,<1",
  "pypdf>=5.4,<6",
]

[project.optional-dependencies]
train = [
  "trl>=1.1.0,<2",
  "peft>=0.18,<1",
  "datasets>=4.0,<5",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
